+++ Common/src/main/java/ca/lukegrahamlandry/lib/base/json/BlockPosTypeAdapter.java
@@ -12,22 +12,23 @@ package ca.lukegrahamlandry.lib.base.json;
 import com.google.gson.*;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Vec3i;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.TagParser;
 
 import java.lang.reflect.Type;
 
-public class BlockPosTypeAdapter implements JsonDeserializer<BlockPos>, JsonSerializer<BlockPos> {
+public class BlockPosTypeAdapter implements JsonDeserializer<Vec3i>, JsonSerializer<Vec3i> {
     public BlockPosTypeAdapter() {
     }
 
-    public BlockPos deserialize(JsonElement data, Type type, JsonDeserializationContext ctx) throws JsonParseException {
+    public Vec3i deserialize(JsonElement data, Type type, JsonDeserializationContext ctx) throws JsonParseException {
         String values = data.getAsString();
         String[] parts = values.split(",");
         return new BlockPos(Integer.parseInt(parts[0]), Integer.parseInt(parts[1]), Integer.parseInt(parts[2]));
     }
 
-    public JsonElement serialize(BlockPos obj, Type type, JsonSerializationContext ctx) {
+    public JsonElement serialize(Vec3i obj, Type type, JsonSerializationContext ctx) {
         String values = obj.getX() + "," + obj.getY() + "," + obj.getZ();
         return new JsonPrimitive(values);
     }