+++ Common/src/main/java/ca/lukegrahamlandry/lib/base/json/JsonHelper.java
@@ -14,6 +14,7 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.TypeAdapterFactory;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Vec3i;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
@@ -28,6 +29,7 @@ public class JsonHelper {
             .registerTypeAdapter(CompoundTag.class, new NbtTypeAdapter())
             .registerTypeAdapter(ItemStack.class, new ItemStackTypeAdapter())
             .registerTypeAdapter(BlockPos.class, new BlockPosTypeAdapter())
+            .registerTypeAdapter(Vec3i.class, new BlockPosTypeAdapter())
             .registerTypeAdapterFactory(new RegistryObjectTypeAdapterFactory());
 
     private static Gson GSON = GSON_BUILDER.create();