+++ Common/src/main/java/ca/lukegrahamlandry/lib/config/ConfigWrapper.java
@@ -11,19 +11,16 @@ package ca.lukegrahamlandry.lib.config;
 
 import ca.lukegrahamlandry.lib.base.Available;
 import ca.lukegrahamlandry.lib.base.InternalUseOnly;
+import ca.lukegrahamlandry.lib.base.WrappedData;
 import ca.lukegrahamlandry.lib.base.json.JsonHelper;
-import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.level.storage.LevelResource;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.Reader;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Type;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -31,7 +28,7 @@ import java.nio.file.StandardCopyOption;
 import java.util.*;
 import java.util.function.Supplier;
 
-public class ConfigWrapper<T> implements Supplier<T> {
+public class ConfigWrapper<T> extends WrappedData<T, ConfigWrapper<T>> implements Supplier<T> {
     /**
      * Creates a new config object for reading settings from player editable files.
      * The config will be synced to all clients, so it may be used from common code.
@@ -78,7 +75,7 @@ public class ConfigWrapper<T> implements Supplier<T> {
      */
     public ConfigWrapper<T> named(String name){
         this.name = JsonHelper.safeFileName(name);
-        this.createLogger();
+        this.updateLogger();
         return this;
     }
 
@@ -87,7 +84,7 @@ public class ConfigWrapper<T> implements Supplier<T> {
      */
     public ConfigWrapper<T> dir(String subDirectory){
         this.subDirectory = JsonHelper.safeFileName(subDirectory);
-        this.createLogger();
+        this.updateLogger();
         return this;
     }
 
@@ -114,22 +111,12 @@ public class ConfigWrapper<T> implements Supplier<T> {
     }
 
     /**
-     * Set the gson instance that will be used for config serialization/deserialization when interacting with files or the network.
-     * This allows you to register your own type adapters. See JsonHelper for defaults provided.
-     * GsonBuilder#setPrettyPrinting will automatically be called when writing defaults to a file (but not for sending over network).
-     */
-    public ConfigWrapper<T> withGson(Gson gson){
-        this.gson = gson;
-        return this;
-    }
-
-    /**
      * Instead of reading one of the specified object from the config file, read a list of them.
      * @param <L> {@code List<T>}
      */
     public <L extends List<T>> ConfigWrapper<L> listOf(){
         ALL.remove(this);
-        TypeToken<L> type = (TypeToken<L>) TypeToken.getParameterized(ArrayList.class, this.clazz);
+        TypeToken<L> type = (TypeToken<L>) TypeToken.getParameterized(ArrayList.class, this.getValueClass());
         ConfigWrapper<L> newWrapper = new ConfigWrapper<>(type, this.side);
         return newWrapper.withSettings(this);
     }
@@ -151,18 +138,9 @@ public class ConfigWrapper<T> implements Supplier<T> {
      */
     public <K, M extends Map<K, T>> ConfigWrapper<M> mapOf(Class<K> keyClass){
         ALL.remove(this);
-        TypeToken<M> type = (TypeToken<M>) TypeToken.getParameterized(HashMap.class, keyClass, this.clazz);
+        TypeToken<M> type = (TypeToken<M>) TypeToken.getParameterized(HashMap.class, keyClass, this.getValueClass());
         ConfigWrapper<M> newWrapper = new ConfigWrapper<>(type, this.side);
         return newWrapper.withSettings(this);
-    }
-
-    /**
-     * The default value will always be automatically set to the parameterless constructor of T (and initialization will fail if such a constructor is not available).
-     * This method allows you to call ConfigWrapper#listOf but not default to an empty list.
-     */
-    public ConfigWrapper<T> setDefaultValue(Supplier<T> v){
-        this.defaultValue = v;
-        return this;
     }
 
     public ConfigWrapper<T> onLoad(Runnable action){
@@ -179,7 +157,7 @@ public class ConfigWrapper<T> implements Supplier<T> {
     public T get() {
         if (!this.loaded) {
             if (this.side == Side.CLIENT) this.load();
-            else this.logger.info("reading config before calling ConfigWrapper#load, default values will be used for now");
+            else this.getLogger().info("reading config before calling ConfigWrapper#load, default values will be used for now");
         }
         return this.value;
     }
@@ -201,7 +179,7 @@ public class ConfigWrapper<T> implements Supplier<T> {
      * Requires Packets module
      */
     public void sync() {
-        if (this.side != Side.SYNCED) this.logger.error("called ConfigWrapper#sync but side=" + this.side + ". Ignoring.");
+        if (this.side != Side.SYNCED) this.getLogger().error("called ConfigWrapper#sync but side=" + this.side + ". Ignoring.");
         else new ConfigSyncMessage(this).sendToAllClients();
     }
 
@@ -211,18 +189,18 @@ public class ConfigWrapper<T> implements Supplier<T> {
      */
     public void load(){
         if (this.side.inWorldDir && server == null) {
-            this.logger.error("cannot load server config before server init. default values will be used for now");
+            this.getLogger().error("cannot load server config before server init. default values will be used for now");
             return;
         }
         if (!Files.exists(this.getFilePath())) this.writeDefaultFile();
 
         try {
             Reader reader = Files.newBufferedReader(this.getFilePath());
-            this.value = this.getGson().fromJson(reader, this.actualType);
+            this.value = this.getGson().fromJson(reader, this.getValueClass());
             reader.close();
-            this.logger.info("config loaded from " + this.displayPath());
+            this.getLogger().info("config loaded from " + this.displayPath());
         } catch (IOException e) {
-            this.logger.error("failed to load config from " + this.displayPath());
+            this.getLogger().error("failed to load config from " + this.displayPath());
             e.printStackTrace();
         }
 
@@ -232,18 +210,14 @@ public class ConfigWrapper<T> implements Supplier<T> {
 
     public static MinecraftServer server;
     public static List<ConfigWrapper<?>> ALL = new ArrayList<>();
-    public Supplier<T> defaultValue;
     private String name;
     public final Side side;
     public boolean shouldReload = true;
-    public Class<T> clazz;
     protected T value;
     private Logger logger;
     private boolean loaded = false;
     private String fileExtension;
-    private Gson gson;
     private String subDirectory = null;
-    Type actualType;
     private Runnable onLoadAction = () -> {};
 
     private ConfigWrapper(Class<T> clazz, Side side){
@@ -252,25 +226,12 @@ public class ConfigWrapper<T> implements Supplier<T> {
     }
 
     public ConfigWrapper(TypeToken<T> type, Side side){
+        super(type);
         this.side = side;
         this.fileExtension = "json5";
-        this.withGson(JsonHelper.get());
         ALL.add(this);
-
-        this.actualType = type.getType();
-        this.clazz = (Class<T>) type.getRawType();
-        this.defaultValue = this::getDefaultInstance;
-        this.value = this.defaultValue.get();
+        this.value = this.getDefaultValue();
         this.named(type.toString());
-    }
-
-    private T getDefaultInstance(){
-        try {
-            return this.clazz.getConstructor().newInstance();
-        } catch (NoSuchMethodException | InvocationTargetException | InstantiationException | IllegalAccessException e) {
-            this.logger.error(clazz.getName() + " does not have a public parameterless constructor");
-            throw new RuntimeException(clazz.getName() + " does not have a public parameterless constructor", e);
-        }
     }
 
     private static String defaultName(Class<?> clazz){
@@ -294,10 +255,10 @@ public class ConfigWrapper<T> implements Supplier<T> {
                 if (Files.exists(defaultLocation)){
                     try {
                         Files.copy(defaultLocation, this.getFilePath(), StandardCopyOption.REPLACE_EXISTING);
-                        this.logger.info("found global default config " + defaultLocation.toAbsolutePath().toFile().getCanonicalPath());
+                        this.getLogger().info("found global default config " + defaultLocation.toAbsolutePath().toFile().getCanonicalPath());
                         return;
                     } catch (IOException e){
-                        this.logger.error("global instance config file existed but could not be copied. generating default");
+                        this.getLogger().error("global instance config file existed but could not be copied. generating default");
                         e.printStackTrace();
                     }
                 }
@@ -305,11 +266,11 @@ public class ConfigWrapper<T> implements Supplier<T> {
         }
 
         try {
-            String configData = GenerateComments.commentedJson(this.defaultValue.get(), this.getGson());
+            String configData = GenerateComments.commentedJson(this.getDefaultValue(), this.getGson());
             Files.write(this.getFilePath(), configData.getBytes());
-            this.logger.info("wrote default config to " + this.displayPath());
+            this.getLogger().info("wrote default config to " + this.displayPath());
         } catch (IOException e){
-            this.logger.error("failed to write default config to " + this.displayPath());
+            this.getLogger().error("failed to write default config to " + this.displayPath());
             e.printStackTrace();
         }
     }
@@ -342,15 +303,12 @@ public class ConfigWrapper<T> implements Supplier<T> {
         }
     }
 
-    public Gson getGson(){
-        return this.gson == null ? JsonHelper.get() : this.gson;
-    }
-
-    private void createLogger(){
-        String id = ConfigWrapper.class.getName() + ": ";
+    @Override
+    protected String getAdditionalLoggerId() {
+        String id = "";
         if (this.getSubDirectory() != null) id = id + this.getSubDirectory() + "/";
         id += this.getName() + "-" + side.name();
-        this.logger = LoggerFactory.getLogger(id);
+        return id;
     }
 
     public String getName() {
@@ -367,7 +325,7 @@ public class ConfigWrapper<T> implements Supplier<T> {
         this.ext(other.fileExtension);
         this.withGson(other.getGson());
         this.shouldReload = other.shouldReload;
-        if (other.clazz == this.clazz) this.setDefaultValue((Supplier<T>) other.defaultValue);
+        if (other.valueType.equals(this.valueType)) this.setDefaultValue(() -> (T) other.getDefaultValue());
         return this;
     }