+++ Common/src/main/java/ca/lukegrahamlandry/lib/data/DataWrapper.java
@@ -10,19 +10,18 @@
 package ca.lukegrahamlandry.lib.data;
 
 import ca.lukegrahamlandry.lib.base.Available;
+import ca.lukegrahamlandry.lib.base.WrappedData;
 import ca.lukegrahamlandry.lib.base.json.JsonHelper;
-import ca.lukegrahamlandry.lib.config.ConfigWrapper;
 import ca.lukegrahamlandry.lib.data.impl.GlobalDataWrapper;
 import ca.lukegrahamlandry.lib.data.impl.LevelDataWrapper;
 import ca.lukegrahamlandry.lib.data.impl.PlayerDataWrapper;
-import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
@@ -34,14 +33,13 @@ import java.util.Locale;
  * Same for fields of T, they can't extend the type of the field, they must be GenericHolders as well.
  * Instead of using holders, you could write your own type adapter that saves the exact type info and call setGson.
  */
-public abstract class DataWrapper<T> {
-
+public abstract class DataWrapper<T, S extends DataWrapper<T, S>> extends WrappedData<T, S> {
     /**
      * Creates a DataWrapper tracking one object per server.
      * @param clazz the type of object to be saved. Must have a public constructor that takes no parameters to create the default value.
      */
     public static <T> GlobalDataWrapper<T> global(Class<T> clazz){
-        return new GlobalDataWrapper<>(clazz);
+        return new GlobalDataWrapper<>(TypeToken.get(clazz));
     }
 
     /**
@@ -49,7 +47,7 @@ public abstract class DataWrapper<T> {
      * @param clazz the type of object to be saved. Must have a public constructor that takes no parameters to create the default value.
      */
     public static <T> LevelDataWrapper<T> level(Class<T> clazz){
-        return new LevelDataWrapper<>(clazz);
+        return new LevelDataWrapper<>(TypeToken.get(clazz));
     }
 
     /**
@@ -57,71 +55,60 @@ public abstract class DataWrapper<T> {
      * @param clazz the type of object to be saved. Must have a public constructor that takes no parameters to create the default value.
      */
     public static <T> PlayerDataWrapper<T> player(Class<T> clazz){
-        return new PlayerDataWrapper<>(clazz);
+        return new PlayerDataWrapper<>(TypeToken.get(clazz));
     }
 
     /**
      * Mark the DataWrapper to be synced to all clients.
      */
-    public <W extends DataWrapper<T>> W synced(){
+    public S synced(){
         if (!Available.NETWORK.get()) throw new RuntimeException("Called DataWrapper#synced but WrapperLib Network module is missing.");
         this.shouldSync = true;
-        return (W) this;
+        return (S) this;
     }
 
     /**
      * Mark the DataWrapper to be saved to disk with the world.
      */
-    public <W extends DataWrapper<T>> W saved(){
+    public S saved(){
         this.shouldSave = true;
-        return (W) this;
+        return (S) this;
     }
 
     /**
      * Set the location to be used for your data file.
      * If saved, the file will be [namespace]/[path]-[side].[ext]
      */
-    public <W extends DataWrapper<T>> W named(ResourceLocation name){
+    public S named(ResourceLocation name){
         this.dir(name.getNamespace());
         this.named(name.getPath());
-        return (W) this;
+        return (S) this;
     }
 
     /**
      * @param name the name of the DataWrapper. This will be used for the filename if saved and for matching instances when syncing.
      */
-    public <W extends DataWrapper<T>> W named(String name){
+    public S named(String name){
         this.name = JsonHelper.safeFileName(name);
-        this.createLogger();
-        return (W) this;
+        this.updateLogger();
+        return (S) this;
     }
 
     /**
      * @param subDirectory the category name of the DataWrapper. This will be used as the folder if saved and for matching instances when syncing.
      */
-    public <W extends DataWrapper<T>> W dir(String subDirectory){
+    public S dir(String subDirectory){
         this.subDirectory = JsonHelper.safeFileName(subDirectory);
-        this.createLogger();
-        return (W) this;
+        this.updateLogger();
+        return (S) this;
     }
 
     /**
      * @param fileExtension the file extension to be used when writing to disk.
      */
-    public <W extends DataWrapper<T>> W ext(String fileExtension){
+    public S ext(String fileExtension){
         this.fileExtension = fileExtension;
-        return (W) this;
-    }
-
-    /**
-     * Set the gson instance that will be used for data serialization/deserialization when interacting with files or the network.
-     * This allows you to register your own type adapters. See JsonHelper for defaults provided.
-     * GsonBuilder#setPrettyPrinting will automatically be called when writing defaults to a file (but not for sending over network).
-     * @param gson the serializer to be used
-     */
-    public <W extends DataWrapper<T>> W withGson(Gson gson){
-        this.gson = gson;
-        return (W) this;
+        return (S) this;
     }
 
     // API
@@ -145,10 +132,9 @@ public abstract class DataWrapper<T> {
 
     // IMPL
 
-    public static List<DataWrapper<?>> ALL = new ArrayList<>();
+    public static List<DataWrapper<?, ?>> ALL = new ArrayList<>();
     public static MinecraftServer server;
 
-    public final Class<T> clazz;
     private String name;
     public String fileExtension = "json";
     protected String subDirectory = null;
@@ -157,12 +143,10 @@ public abstract class DataWrapper<T> {
     protected boolean isLoaded = false;
     protected boolean isDirty = false;
     public Logger logger;
-    private Gson gson;
 
-    protected DataWrapper(Class<T> clazz){
-        this.clazz = clazz;
-        this.named(defaultName(clazz));
-        this.createDefaultInstance();
+    protected DataWrapper(TypeToken<T> type){
+        super(type);
+        this.named(defaultName(type.getRawType()));
         ALL.add(this);
     }
 
@@ -180,29 +164,17 @@ public abstract class DataWrapper<T> {
         return this.subDirectory;
     }
 
-    public Gson getGson(){
-        return this.gson == null ? JsonHelper.get() : this.gson;
-    }
-
     private static String defaultName(Class<?> clazz){
         return clazz.getSimpleName().toLowerCase(Locale.ROOT);
     }
 
-    public T createDefaultInstance() {
-        try {
-            return clazz.getConstructor().newInstance();
-        } catch (NoSuchMethodException | InvocationTargetException | InstantiationException | IllegalAccessException e) {
-            this.logger.error(clazz.getName() + " does not have a public parameterless constructor");
-            throw new RuntimeException(clazz.getName() + " does not have a public parameterless constructor", e);
-        }
-    }
-
     public static final Logger LOGGER = LoggerFactory.getLogger(DataWrapper.class);
-    protected void createLogger(){
-        String id = DataWrapper.class.getName() + ": ";
+
+    protected String getAdditionalLoggerId(){
+        String id = "";
         if (this.getSubDirectory() != null) id = id + this.getSubDirectory() + "/";
         id += this.getName();
-        this.logger = LoggerFactory.getLogger(id);
+        return id;
     }
 
     public static String forDisplay(Path path){