+++ Common/src/main/java/ca/lukegrahamlandry/lib/data/impl/GlobalDataWrapper.java
@@ -14,6 +14,7 @@ import ca.lukegrahamlandry.lib.data.DataWrapper;
 import ca.lukegrahamlandry.lib.data.sync.GlobalDataSyncMessage;
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
+import com.google.gson.reflect.TypeToken;
 import net.minecraft.world.level.storage.LevelResource;
 
 import java.io.IOException;
@@ -22,16 +23,16 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.function.Supplier;
 
-public class GlobalDataWrapper<T> extends DataWrapper<T> implements Supplier<T> {
+public class GlobalDataWrapper<T> extends DataWrapper<T, GlobalDataWrapper<T>> implements Supplier<T> {
     T value;
-    public GlobalDataWrapper(Class<T> clazz) {
-        super(clazz);
+    public GlobalDataWrapper(TypeToken<T> type) {
+        super(type);
     }
 
     @Override
     public T get() {
         if (!this.isLoaded) {
-            this.logger.error("cannot call DataWrapper#get (a) before server startup (b) on client if unsynced");
+            this.getLogger().error("cannot call DataWrapper#get (a) before server startup (b) on client if unsynced");
             return null;
         }
 
@@ -42,7 +43,7 @@ public class GlobalDataWrapper<T> extends DataWrapper<T> implements Supplier<T> 
      * Resets the data to default values.
      */
     public void clear(){
-        this.value = this.createDefaultInstance();
+        this.value = this.getDefaultValue();
         this.setDirty();
     }
 
@@ -50,7 +51,7 @@ public class GlobalDataWrapper<T> extends DataWrapper<T> implements Supplier<T> 
     public void load() {
         if (server == null) {
             String msg = "cannot call DataWrapper#load (a) before server startup (b) on client";
-            this.logger.error(msg);
+            this.getLogger().error(msg);
             throw new RuntimeException(msg);
         }
 
@@ -62,11 +63,11 @@ public class GlobalDataWrapper<T> extends DataWrapper<T> implements Supplier<T> 
 
         try {
             Reader reader = Files.newBufferedReader(this.getFilePath());
-            this.value = this.getGson().fromJson(reader, this.clazz);
+            this.value = this.getGson().fromJson(reader, this.getValueClass());
             reader.close();
         } catch (IOException | JsonSyntaxException e) {
             String msg = "failed to load data from " + forDisplay(this.getFilePath());
-            this.logger.error(msg);
+            this.getLogger().error(msg);
             e.printStackTrace();
         }
 
@@ -82,13 +83,13 @@ public class GlobalDataWrapper<T> extends DataWrapper<T> implements Supplier<T> 
         try {
             Files.write(path, json.getBytes());
         } catch (IOException e) {
-            this.logger.error("failed to write data to " + forDisplay(path));
+            this.getLogger().error("failed to write data to " + forDisplay(path));
         }
     }
 
     @Override
     public void sync() {
-        if (!this.shouldSync) this.logger.error("called DataWrapper#sync but shouldSync=false");
+        if (!this.shouldSync) this.getLogger().error("called DataWrapper#sync but shouldSync=false");
         else new GlobalDataSyncMessage(this).sendToAllClients();
     }