+++ Common/src/main/java/ca/lukegrahamlandry/lib/data/impl/file/SingleFileHandler.java
@@ -22,9 +22,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 
 public class SingleFileHandler<K, I, V> implements MapFileHandler<K, I, V> {
-    private final MapDataWrapper<K, I, V> wrapper;
+    private final MapDataWrapper<K, I, V, ?> wrapper;
 
-    public SingleFileHandler(MapDataWrapper<K, I, V> wrapper) {
+    public SingleFileHandler(MapDataWrapper<K, I, V, ?> wrapper) {
         this.wrapper = wrapper;
     }
 
@@ -36,7 +36,7 @@ public class SingleFileHandler<K, I, V> implements MapFileHandler<K, I, V> {
         try {
             Files.write(this.getFilePath(), json.getBytes());
         } catch (IOException e) {
-            this.wrapper.logger.error("failed to write data to " + DataWrapper.forDisplay(this.getFilePath()));
+            this.wrapper.getLogger().error("failed to write data to " + DataWrapper.forDisplay(this.getFilePath()));
         }
     }
 
@@ -52,7 +52,7 @@ public class SingleFileHandler<K, I, V> implements MapFileHandler<K, I, V> {
             this.wrapper.loadFromMap(fileInfo);
         } catch (IOException | JsonSyntaxException e) {
             String msg = "failed to load data from " + DataWrapper.forDisplay(this.getFilePath());
-            this.wrapper.logger.error(msg);
+            this.wrapper.getLogger().error(msg);
             e.printStackTrace();
         }
     }