+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 - [ ] quilt example mod
 - [ ] config array / map comments on fields brought to top if list empty
+- [ ] split synced data pack into multiple packets if it gets too big 
 - [X] world configs check ./config (as well as ./defaultconfigs) before generating default
 
 ## 1.1.0